Splitting Modpacks + MyMods. Re-added packages.config for auto-pull from nuget at compile! Added Fody as nuget reference. Set compile target icon.
