+++ Properties/Resources.resx
@@ -127,6 +127,9 @@
   <data name="greenDot_15px" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\greenDot_15px.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
+  <data name="HaloHelmetIcon" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\HaloHelmetIcon.ico;System.Drawing.Icon, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
   <data name="HaloHelmetIcon_small" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\HaloHelmetIcon_small.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
