+++ packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Costura.Fody" version="4.1.0" targetFramework="net472" />
+  <package id="Fody" version="6.0.0" targetFramework="net472" developmentDependency="true" />
+  <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net472" />
+  <package id="System.IO.Compression" version="4.3.0" targetFramework="net472" />
+</packages>
